Using common library to check for job output and see if it matches fi… (#28)
Implementing 5.4 of FHIR spec so that users can request a file download using the output they received in the previous API status call. Using common library to check for job output and see if it matches files, so that file can be downloaded and checked for correct output. Adding property for environment variable that can be accessed to determine location on filesystem of files for download.