Feature/ab2d 331 parse hpms file (#32)
Parse the sponsors and contract file so that records in the database can be created. This operation is accessed by an admin API call, which will be locked down in the future. The operation is idempotent, so this should be able to be ran over and over again. Adding support for both XLS and XLSX file types.