Feature/ab2d 76 (#42)
Adding authentication via JWT and okta. The JWT is verified every request the user makes, a filter exception handler is added too, in order to handle exceptions that come from the authentication filter. Okta code is called in order to verify the JWT that is sent, and then there are checks into the user_account table to make sure that the user is valid. Updated tests to add the Authorization header, and created a self signed JWT to make tests more robust.