Feature/ab2d 403 test containers (#77)
Adding containers so that the tests run in a Postgres environment. This will require using the Testcontainers annotation on a test class, as well as a variable annotated with Container, so that the container will run for a test class. The container will be shared for tests running in the same class, and be brought up again for different modules.