Feature/ab2d 610 endpoint (#82)
Adding in an endpoint to kick off a job based on contract number, adding some testing utilities, adding a migration so that contract can be tracked in the job table if it's used. Put in checks to ensure that a contract is attested on and connected to the user's sponsor.