Feature/ab2d 673 automated test v2 (#110)
e2e tests that make several API requests, and then allow the worker to process the data. After the worker has processed the data verify that the data is correct. Adds a new docker compose so that the application can be built with a different environment and allow e2e specific variables to take place and share a volume between worker and API.