Setting max concurrent jobs to 3 (#141)
Add a column in the user_account table that will set the maximum concurrent jobs a user can have to 1 by default, and adding a migration that will change it to 3 for all users. The exports will enforce that a user can only have a max number of jobs to whatever their limit is set to.