Feature/ab2d 785 load testing (#140)
Refactoring code into new class APIClient so that it can be reusable, adding JMeter tests that can be run from a JMeter command line or GUI, load tests will keep track of each request's time that it took. The tests are configurable so that they can hit different environments. Instructions are provided on how to run the tests from JMeter.