If the file can be connected to a job, retrieve the job so that the d… (#158)
If the file can be connected to a job, retrieve the job so that the deletion can factor in whether or not the job has expired yet. The completion date of the job will be used to determine if the job should be deleted or not. If the file is not connected to a job, use the existing logic.