Feature/ab2d 1029 create contract adapter impl (#190)
Create a new class ContractAdapterImpl to implement the ContractAdapter interface.
This new implementation enables the worker to make a call out to BFD via the operations added to the BFDClientImpl dependency to fetch the active beneficiaries in a given contract. The new BFD api does provides data only for a given month. So add logic to call the new BFD api multiple times,  once for each month in the current year from January till the current month.

* Add Junit tests
* Address review comments