Feature/ab2d 1033 load properties (#193)
Create a migration to create a table for storing properties and insert 3 that will be able to be reloaded. On application startup the properties will be read from the database and then set so that places throughout the application that access properties can use the conventional way of accessing them in Spring.