Feature/ab2d 1058 switch (#216)
Add a properties value that determines whether or not we are in maintenance mode. This property is stored in the database and is checked by the API when a request comes in to create a job. If we are in maintenance mode, no new jobs can be created, but users can still download/delete/get the status of jobs.