Feature/ab2d 1146 status file response (#229)
Create a migration so that checksum and file length columns are added to the database. Set the checksum and file length from the worker when the job is completed so that when a user makes a Bulk Status call that has a 200 status, the checksum and file length will be displayed to the user.