Feature/ab2d 1211 e2e opt out (#269)
Add an e2e test for opt out that will check to see if a user that has opted out remains in the data download. Add environment variables for connecting to AWS and update the BFD hash so that it will be able to work with the BFD sandbox we are connecting to.