Feature/ab2d 1416 replace contract adapter (#305)
Change the ContractAdapterImpl to be the primary one being used so that it will grab data from BFD. Update all contracts in the system so that they start with a Z instead of an S. Change the tests that don't use mocks so that they use the correct contracts. Add a variable so that we can skip billable period checks since the billable period is not up to date with the dates under contract.