apply request logging filter only to public requests
Filters are loaded as regular expressions that are compiled and converted to a single Predicate. If a request uri matches any regex then the Predicate returns false, meaning the request should not be logged.

The regexes are provided as a property which may be modified at startup but is defaulted.