Feature/ab2d 1944 limit logged requests (#418)
* apply request logging filter to limit logging of public requests

Filters are loaded as regular expressions that are compiled and converted to a single Predicate. If a request uri matches any regex then the Predicate returns false, meaning the request should not be logged.

The regexes are provided as a property which may be modified at startup but is defaulted.

* filter out health api

Have not written tests but some original tests are failing.

* suppress unused warning

* provide more comprehensive tests

Use ReflectionTestUtils to change the value of the uriFilters to facilitate tests.

Change the default URI filter to only filter out the /health api and no other apis. Similar URIs that would not be filtered /job/health, /healthy, /heal, etc.

* clear residual events from container

* comment that commas split argument lists

Co-authored-by: wnyffenegger <will.nyffenegger@semanticbits.com>