refactor update logic for coverage into a separate driver
The goal of the refactor is to keep the coverage processor as dumb as possible to limit complexity. By separating the logic for deciding which coverage periods to load metadata for and the actual loading of metadata we keep the ability to create a separate service in the future if necessary.