Feature/ab2d 2742 generalize contract search lock (#543)
* add mechanism to detect and add new coverage periods

includes component to specify the attestation date of a contract as EST instead of current time zone.

* coverage search quartz job created and configured

refactored coverage search to separate package

* handle rebase

* allow toggling both discovery and queueing of coverage periods

* use zone id constant instead of zone string

* rename method and add comment

* refactor update logic for coverage into a separate driver

The goal of the refactor is to keep the coverage processor as dumb as possible to limit complexity. By separating the logic for deciding which coverage periods to load metadata for and the actual loading of metadata we keep the ability to create a separate service in the future if necessary.

* refactor update logic for coverage into a separate driver

The goal of the refactor is to keep the coverage processor as dumb as possible to limit complexity. By separating the logic for deciding which coverage periods to load metadata for and the actual loading of metadata we keep the ability to create a separate service in the future if necessary.

* prioritize coverage searches for existing eob jobs

* refactor update logic for coverage into a separate driver

The goal of the refactor is to keep the coverage processor as dumb as possible to limit complexity. By separating the logic for deciding which coverage periods to load metadata for and the actual loading of metadata we keep the ability to create a separate service in the future if necessary.

* minor rebase modifications

* clean up merge

* update to use constant for epoch instead of function

* address PR comments and extend lock ttl to avoid issues

JdbcLockRegistry will destroy a lock if that lock has not been touched for ttl seconds. The default is 10 seconds which is much too short so we have extended it to 10 minutes.

* make CoverageDriverException a runtime exception