PLT-746 Fix if statement in "promote" workflow (#1415)
## 🎫 Ticket

[PLT-746](https://jira.cms.gov/browse/PLT-746)

## 🛠 Changes

"fi" was added at the end of set environment variable statement

## ℹ️ Context

ab2d team is trying to deploy some code to "prod-test" and looks like
there are errors with existing code syntax in the. "promote" work flow .

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Link to a successful manually trigger
https://github.com/CMSgov/ab2d/actions/runs/11666218254

---------

Co-authored-by: Maboh Christopher <chris.maboh@aquia.io>