Add details for 'coverage period missing or year,month query incorrect' log statements (#1418)
## 🎫 Ticket

- https://jira.cms.gov/browse/AB2D-6418
- https://jira.cms.gov/browse/AB2D-6425

## 🛠 Changes

Add details to log message for `coverage period missing or year,month
query incorrect` error messages.

## ℹ️ Context

When we receive a Slack notification of a failed job, the relevant
details (e.g. coverage month and year) are not present in Splunk. This
PR adds those details for debugging purposes.

![image](https://github.com/user-attachments/assets/87cfcea5-c673-4198-a65d-fb6315cb5f9b)

## 🧪 Validation

Added unit tests that verified additional statements were printed to the
logs

![image](https://github.com/user-attachments/assets/b7b29df6-7cf8-46e7-8128-846537ed2a99)