AB2D-6276 Research LOE for HttpGet to HttpPost request change (#1424)
## 🎫 Ticket

https://jira.cms.gov/browse/AB2D-6276


## 🛠 Changes

Increased bfd version from 2.3.0 to 2.4.0

## ℹ️ Context

Context: BFD will stop supporting "GET" requests soon , so we need to
update all GET requests to post

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->