Make sure prod and prodtest use the same account (#1452)
## 🎫 Ticket

Adhoc change

## 🛠 Changes

This is to address failures on run where prodtest is selected but the
account number remains blank.

## ℹ️ Context

In our step we use the environment chosen for the account number.
However prod and prod-test use the same account number. We have to
explicitly say they are using the same account number.
## 🧪 Validation
https://github.com/CMSgov/ab2d/actions/runs/13970690678/job/39111416664

the assume steps works fine in the run above. It failed on the backend
because we do not deploy api in prod-test

---------

Co-authored-by: oluwolenpbc <boluwole@brians-MacBook-Pro-4.local>