PLT-1038 update idm url (#1469)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1038

## 🛠 Changes

Update the okta url for idm

## ℹ️ Context


https://cmsgov.slack.com/archives/CNVRZ73NF/p1744121296051429?thread_ts=1744054808.021109&cid=CNVRZ73NF

We are having issues with fips mode being enforced and idm not allowing
access from our runners because of this.

We want to change the vanity url we're accessing so that we don't run
into this issue.

## Security

IDM by default doesn't require fips with TLS 1.2. Amazon Linux 2023 uses
openssl 3.x.x by default and with this comes with TLS 1.3 and fips
enabled. By switching the vanity URL we access we won't hit the same
issue

## 🧪 Validation
This has been run and works now
https://github.com/CMSgov/ab2d/actions/runs/14738998143/job/41372081943

Co-authored-by: oluwolenpbc <boluwole@brians-MacBook-Pro-4.local>