PLT-1099: Introduce SOPS-enabled SSM Configuration (#1485)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1099

## 🛠 Changes
While this is a proof-of-concept for using a SOPS-enabled tf module,
it's still effective. This introduces ops/services hierarchy using
opentofu as the infrastructure-as-code binary. CI and other amenities
are forthcoming, but today, this manages the majority of the SSM-stored
configuration values in greenfield dev.

## ℹ️ Context

greenfield migration

## 🧪 Validation
This has successfully deployed against the greenfield dev environment.

---------

Co-authored-by: oluwolenpbc <boluwole@brians-MacBook-Pro-4.local>