AB2D-6707: Fix, Simplify, Improve Selective GHA (#1491)
## 🎫 Ticket

https://jira.cms.gov/browse/AB2D-6707

## 🛠 Changes
The below is the one-line fix for the impacted workflows:
```diff
-    if: needs.workflow.output.files
+    if: needs.workflow.outputs.files
```

In additional to addressing the above typo, I've extended the regular
expression to also trigger workflows on changes to the logic in the
workflows themselves, simplified the logic in the checking job, and
updated the comment with relevant GitHub community discussions on this
workaround (hack).

## ℹ️ Context
Improvements to efficiency, the developer experience, cost, and
maintainability.

## 🧪 Validation
- Workflows on **this** change set are requiring and executing the
workflows as expected.
- I created subsequent PR (#1492) that was branched from **this** change
set to demonstrate that workflows **do not** execute as desired.