AB2D-6755: Support Ephemeral Environments (#1530)
## 🎫 Ticket

https://jira.cms.gov/browse/AB2D-6755

## 🛠 Changes
Between Aurora and Ephemeral Environment Support, these are a summary of
the high notes:
* Prefer Aurora Clusters in parent/established `dev`, `test`, and
`sandbox` environments
* Support both RDS DB Instances and Aurora Clusters by parent
environment (temporarily supporting `prod` until migration date 🤞)
* In conjunction with ongoing PLT-1099 work, support copy of subset of
enduring/parent SSM parameters into ephemera as part of `10-config`
* Further support PLT-1077 and simplify network log access configuration
for load balancers
* Re-use EFS, S3, and ACM enduring environment resources as appropriate
for ephemera
* Defer to same environment `local.env` for SSM parameter lookup instead
of parent environment `local.parent_env` (e.g. ephemeral environments
have their own SSM hierarchies, and specific permissions for those
hierarchies)
* Default to standard Aurora Storage in temporary ds (data store) tf
module
* Implement stricter `require` in JDBC URIs


## ℹ️ Context
While this generally introduces support for ephemeral environments in
AB2D, this was executed as part of and in support of a migration to
Aurora Clusters. As a result, the majority of changes are here to
support ephemeral environments with other changes here that are
necessary to extend support for Aurora clusters into both ephemeral and
enduring environments.


## 🧪 Validation
* successfully deployed against `dev` without any issues, followed by
e2e tests
* this was used in two ephemeral environments (`6755-dev` and
`6755-prod`, the latter with some modest changes that are not reflected
in **this** change set) supporting the preparation for the aurora
migration tomorrow, 2025-07-17
* PLT-1077 changes that address the load balancer logging are
successfully arriving in splunk