AB2D-6801: Manage optout lambda infrastructure (#1532)
## 🎫 Ticket

https://jira.cms.gov/browse/AB2D-6801

## 🛠 Changes
This is primarily a lift-and-shift from the CDAP-managed
[opt-out-export](https://github.com/CMSgov/cdap/tree/main/terraform/services/opt-out-export)
and
[opt-out-import](https://github.com/CMSgov/cdap/tree/main/terraform/services/opt-out-import)
infrastructure to help more holistically manage AB2D change over time
and bring clarity to the infrastructure resources that support the
beneficiary preferences and opt-out functionality.

This continues the pattern in-use by the majority of the lambda
functions in `30-lambda` where zip archives are published (delivered) to
artifactory and consumed by `30-lambda` for deployment.

## ℹ️ Context
As part of the greenfield migration and more recent aurora migration,
it's clear that AB2D needs to establish more independence and
self-reliance to better control changes to the infrastructure they
ostensibly manage. To this end, resources that support the AB2D services
need to be managed in a standard, arguably simpler way.


### Future Work
* Unused resources formerly managed by the opt-out-export and
opt-out-import modules above as part of `test` and `prod` environments
need to be destroyed, e.g. distinct KMS CMKs and S3 buckets.
* Unused, unmaintained resources from the aforementioned modules in the
`dev` and `sandbox` environments need to be destroyed.
* Overall, more resources in `30-lambda` need to be covered by the
environmental KMS CMK, including the lambda's environment variables, SNS
topics, and SQS queues
* The shifted IAM resources use outdated customer in-line policies where
a customer managed policy/policies would be more auspicious
* The shifted IAM resources are over privileged and need to have more
targeting in the permission grants and accessible resources
* The import/export functions could _probably_ use the same IAM role
instead of distinct IAM roles–they're virtually identical

## 🧪 Validation
As part of the lift and shift, `test` and `prod` resources were
imported. Changes implemented as part of this module were successfully
applied to the newly imported resources.