AB2D-6378: Add script for inferno tests (#1536)
## 🎫 Ticket

https://jira.cms.gov/browse/AB2D-6378

## 🛠 Changes

Add a script to be able to run the inferno test suite with a single
command

## ℹ️ Context

Before this PR, running the inferno tests was a manual process. Now we
can run them with a single command.

This is a step towards ultimately running the inferno conformance tests
as part of an automated process.

## 🧪 Validation

Try running the inferno conformance test suite on the sandbox env or
your local env:

1. Set the Environment Variables:
- `BULK_URL` - the AB2D Endpoint you are trying to hit. ex:
`'https://sandbox.ab2d.cms.gov/api/v2/fhir/'` for AB2D sandbox
- `TOKEN_URL` - the full AB2D Auth endpoint used to get an access token
  - `CLIENT_ID` - Client ID used for authenticating
  - `CLIENT_SECRET` - Client Secret used for authenticating

2. Run the script: `./scripts/run-fhir-tests.sh`

3. Validate that it runs all the way through and prints out the results,
like:

```sh
SUMMARY:
 - Tests Passed: 7
 - Tests Failed: 5
 ```