PLT-521: Move AB2D to BCDA on-call org (#1537)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-521

## 🛠 Changes

cloudwatch sns subscription for splunk on-call in Greenfield account was
added

## ℹ️ Context

Currently, the BCDA, DPC, and Platform teams are in the BCDA on-call org
while AB2D is in its own "CMS-AB2D" org, which makes escalation of
alerts from AB2D to Platform difficult. Therefore the need to migrate
AB2D team to BCDA on-call org

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation
* successfully applied against `sandbox` and `prod`
* email-based subscriptions to VictorOps/Splunk On Call have been
confirmed
<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->

---------

Co-authored-by: Maboh Christopher <chris.maboh@aquia.io>