AB2D-6797: Enable read-only root filesystem (#1535)
## 🎫 Ticket

https://jira.cms.gov/browse/AB2D-6797

## 🛠 Changes

Enable read-only root filesystem

## ℹ️ Context

Set the container read-only root filesystem to true- address security
alerts related to "This control checks if ECS containers are limited to
read-only access to mounted root filesystems. This control fails if the
ReadonlyRootFilesystem parameter in the container definition of ECS task
definitions is set to ‘false’."

## 🧪 Validation

We can see in ECS that the readonlyRootFilesystem is set to true. We can
also see from our exclusions that we see logs in cloudwatch and newrelic
is working. No errors in the logs due to the readonlyRootFilesystem.

---------

Co-authored-by: oluwolenpbc <boluwole@brians-MacBook-Pro-4.local>
Co-authored-by: oluwolenpbc <127983913+oluwolenpbc@users.noreply.github.com>
Co-authored-by: Michael J Burling <michael.burling@cms.hhs.gov>