PLT-1342: drop default force_ssl  (#1620)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1342

## 🛠 Changes

dropping default force_ssl 

## ℹ️ Context

default force_ssl cluster parameter is now being set in aurora module in
cdap repo and there for the need to drop it here

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

The Terraform plan for the Aurora module changes has been validated in
[BCDA-OPS PR #1268](https://github.com/CMSgov/bcda-ops/pull/1268).

---------

Co-authored-by: Maboh Christopher <chris.maboh@aquia.io>