PLT-1374: Enable CloudWatch Log Exports for PostgresQL Logs (#1627)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1374

## 🛠 Changes

Changes the commit hash to reference the latest change to the aurora
module.

## ℹ️ Context

As part of our ongoing SecHub findings remediation, we identified that
Aurora PostgresQL logs were not being exported to CloudWatch. This
change should make those logs available and fix the finding. For more
info on the changes, see this commit:
https://github.com/CMSgov/cdap/commit/231b7d4f1e607b61b58ec497120b6188a998ef1f

## 🧪 Validation
Output verifying the changes after referencing this commit hash:
<details>
<summary>Tofu Plan Output (AB2D dev)</summary>

```
OpenTofu will perform the following actions:

  # module.db.aws_rds_cluster.this will be updated in-place
  ~ resource "aws_rds_cluster" "this" {
      ~ enabled_cloudwatch_logs_exports       = [
          + "postgresql",
        ]
        id                                    = "ab2d-dev"
        tags                                  = {
            "AWS_Backup" = "4hr1dr_d7_w35_m90"
        }
        # (47 unchanged attributes hidden)
    }

Plan: 0 to add, 1 to change, 0 to destroy.
```
</details>

### To be merged after pentest is complete 10/10