[PLT-1406] Uncommenting CloudFront distribution cert alias. (#1631)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1406

## 🛠 Changes

This PR uncomments the AB2D static site TF line that sets the prod
distribution domain alias to 'ab2d.cms.gov'.

## ℹ️ Context

This TF line was commented out during GF deployments because the legacy
account distribution was still using that alias. This must now be
uncommented so that future deployments do not unset that alias.

## 🧪 Validation

The following Tofu plan was run against PROD and shows that the alias is
now unchanged. There does appear to be two instances of drift unrelated
to this change however, where (1) the CloudFront distribution
'PriceClass' value is set to "PriceClass_All" in PROD but to
"PriceClass_100" in TF, and (2) where a bucket policy has been applied
to the origin bucket that does not (or no longer) exists in TF (plan
truncated for clarity):

```
OpenTofu will perform the following actions:

  # data.aws_iam_policy_document.this will be read during apply
  # (depends on a resource or a module with changes pending)
 <= data "aws_iam_policy_document" "this" {
   }

  # aws_cloudfront_distribution.this will be updated in-place
  ~ resource "aws_cloudfront_distribution" "this" {
        id                             = "XXXXXXXXXXXXXX"
      ~ price_class                    = "PriceClass_All" -> "PriceClass_100"
        tags                           = {}
        # (21 unchanged attributes hidden)

        # (5 unchanged blocks hidden)
    }

  # aws_s3_bucket_policy.this will be updated in-place
  ~ resource "aws_s3_bucket_policy" "this" {
    }

Plan: 0 to add, 2 to change, 0 to destroy.
```