AB2D-6940 Add Inferno Commit Hash (#1632)
## 🎫 Ticket

https://jira.cms.gov/browse/AB2D-6940

## 🛠 Changes

pointed the inferno docker build to use a specific commit hash. There is
no functional change right now, since it is still pointing to the
most-recent commit. But now it will stay pointing at this commit, even
when more commits to bulk-data-test-kit are added.

## ℹ️ Context

If we look at the current release, there could be breaking changes, or
vulnerabilities could be introduced intentionally by a bad actor.
Sticking to a specific commit is the more robust solution.

## 🧪 Validation

Make sure FHIR conformance tests still pass


https://github.com/CMSgov/ab2d/actions/runs/18787597185/job/53609664748?pr=1632