PLT-1368 Removing price_class attribute from CloudFront definition. (#1633)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1368

## 🛠 Changes

This PR removes the price_class attribute from the CloudFront resource
definition.

## ℹ️ Context

The static site aws_cloudfront_distribution resource definition was
specifying the price_class attribute with a value of "PriceClass_100".
In order to comply with the CMS Hybrid Cloud Program's CloudFront Strict
Region Routing implementation this attribute will now be omitted so that
the default value of "PriceClass_All" will be picked up going forward.

## 🧪 Validation

A TF plan against the AB2D PROD static site resulted in one unrelated S3
bucket policy change. Given that the static site is already configured
with "PriceClass_All" this plan is as desired in that it will not revert
the value back to "PriceClass_100".