PLT-1447 Security groups for attribution and IDR endpoint (#1653)
## 🎫 Ticket

https://jira.cms.gov/browse/PLT-1447

## 🛠 Changes

Added security groups for attribution and IDR endpoint.

## ℹ️ Context

This change allows for integration with IDR Snowflake for the
attribution service. The attribution security group is created in all
environments and should be used for the attribution service when it is
set up. The idr-endpoint security group is only created in the prod
environment because the PrivateLink endpoint for IDR Snowflake can only
be created in prod.

## 🧪 Validation

I've checked the plan in all environments and applied in prod in order
to create the security group necessary for filing the request with the
CMS Hybrid Cloud team to create the PrivateLink endpoint.